requirements.txt
# React Agent with LangChain RAG Implementation
# Requirements for the ice breaker application with grocery sales data support

# Core LangChain dependencies
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0

# Environment management
python-dotenv>=1.0.0

# Development and code formatting
black>=23.0.0

# Additional dependencies for RAG and data processing
# Uncomment as needed for your specific implementation:
# pandas>=2.0.0              # For data manipulation
# numpy>=1.24.0              # For numerical computing
# faiss-cpu>=1.7.4           # For vector similarity search
# chromadb>=0.4.0            # For vector database
# tiktoken>=0.5.0            # For token counting
# openai>=1.0.0              # Direct OpenAI API access
# python-decouple>=3.8       # Alternative to python-dotenv
# pydantic>=2.0.0            # Data validation

# For enhanced text processing:
# nltk>=3.8                  # Natural language processing
# spacy>=3.7.0               # Advanced NLP
# sentence-transformers>=2.2.0  # Embeddings generation

# For web interface (if needed):
# streamlit>=1.28.0          # Web app framework
# fastapi>=0.104.0           # API framework
# uvicorn>=0.24.0            # ASGI server
